requirements.txt
flask-restplus==0.9.2
Flask-SQLAlchemy==2.1
web3==5.2.2
py-solc==3.2.0
